Changes to the ban checks
Ban evading uses the repository, not the direct DB.
Webhook of ban evading was using the wrong type. Switched to WebhookType.BanEvading
Ban evading now gets logged to console.
banRemainingDuration doesn't overflow if the remaining duration is larger than uint.MaxValue.
When selecting an ID ban, it now selects the one with the longest duration.
Add a check to see if the IP or HWID differ from the previous ban (with an ID ban), and re-ban them for the remaining duration.
Translations for ban evading reason, ban evading admin name, for the re-ban with different IP and/or HWID, and for the commands having no reason specified have been added.
Changed exception when calling `/banhistory` from console from `CommandNotFoundException` to `CommandWrongUsageException`